feat: packages begone!
