--- assets/assets.go
@@ -1,6 +1,0 @@
-package assets
-
-import "embed"
-
-//go:embed data
-var Assets embed.FS
