+++ assets/characters/boss_one.png
