+++ assets/characters/player_right.png
