+++ boss_1.go
@@ -1,26 +1,24 @@
-package entity
+package main
 
 import (
 	"math"
 	"time"
 
-	"github.com/eliasrenman/go-bullet-hell/assets"
-	"github.com/eliasrenman/go-bullet-hell/geometry"
 	"github.com/hajimehoshi/ebiten/v2"
 )
 
-var bossOneImage = assets.LoadImage("characters/boss_one.png", assets.OriginCenter)
+var bossOneImage = LoadImage("characters/boss_one.png", OriginCenter)
 
-func NewBossOne(position geometry.Point) *BossOne {
+func NewBossOne(position Point) *BossOne {
 	e := &Entity{
 		Position: position,
 	}
 	boss := &BossOne{
 		Entity: e,
 		Hitbox: &RectangleHitbox{
-			Size: geometry.Size{Width: 32, Height: 32},
+			Size: Size{Width: 32, Height: 32},
 			BaseHitbox: BaseHitbox{
-				Position: geometry.Vector{X: -16, Y: -16},
+				Position: Vector{X: -16, Y: -16},
 				Owner:    e,
 			},
 		},
@@ -34,7 +32,7 @@ type BossOne struct {
 }
 
 func (boss *BossOne) Draw(screen *ebiten.Image) {
-	bossOneImage.Draw(screen, boss.Position, geometry.Size{Width: 1, Height: 1}, 0)
+	bossOneImage.Draw(screen, boss.Position, Size{Width: 1, Height: 1}, 0)
 }
 
 var schedule = Schedule{
