+++ circlehitbox.go
@@ -1,32 +1,27 @@
-package entity
-
-import (
-	"github.com/eliasrenman/go-bullet-hell/geometry"
-	"github.com/eliasrenman/go-bullet-hell/util"
-)
+package main
 
 type CircleHitbox struct {
 	BaseHitbox
 	Radius float64
 }
 
-func (a *CircleHitbox) CollidesWithCircle(p geometry.Point, b *CircleHitbox) bool {
+func (a *CircleHitbox) CollidesWithCircle(p Point, b *CircleHitbox) bool {
 	aCenter := a.Position.Plus(p)
 	bCenter := b.Position.Plus(b.Owner.Position)
 
 	return aCenter.Distance(bCenter) < a.Radius+b.Radius
 }
 
-func (a *CircleHitbox) CollidesWithRectangle(p geometry.Point, b *RectangleHitbox) bool {
+func (a *CircleHitbox) CollidesWithRectangle(p Point, b *RectangleHitbox) bool {
 	aCenter := a.Position.Plus(p)
 	bMin, bMax := b.GetRectangle(b.Owner.Position)
 
 	// Find the closest point to the circle within the rectangle
-	x := util.ClampFloat(aCenter.X, bMin.X, bMax.X)
-	y := util.ClampFloat(aCenter.Y, bMin.Y, bMax.Y)
+	x := ClampFloat(aCenter.X, bMin.X, bMax.X)
+	y := ClampFloat(aCenter.Y, bMin.Y, bMax.Y)
 
 	// Calculate the distance between the circle's center and this closest point
-	distance := aCenter.Distance(geometry.Point{X: x, Y: y})
+	distance := aCenter.Distance(Point{X: x, Y: y})
 
 	// If the distance is less than the circle's radius, an intersection occurs
 	return distance < a.Radius
@@ -36,7 +31,7 @@ func (hb CircleHitbox) CollidesWith(other Collider) bool {
 	return other.CollidesAt(hb.Owner.Position, &hb)
 }
 
-func (a CircleHitbox) CollidesAt(p geometry.Point, b Collider) bool {
+func (a CircleHitbox) CollidesAt(p Point, b Collider) bool {
 	switch other := b.(type) {
 	case *CircleHitbox:
 		return a.CollidesWithCircle(p, other)
