--- constant/constants.go
@@ -1,8 +1,0 @@
-package constant
-
-import "github.com/eliasrenman/go-bullet-hell/geometry"
-
-var (
-	WORLD_ORIGIN = geometry.Point{X: 0, Y: 0}
-	WORLD_SIZE   = geometry.Size{Width: 350, Height: 550}
-)
