+++ constants.go
@@ -1,0 +1,6 @@
+package main
+
+var (
+	WORLD_ORIGIN = Point{X: 0, Y: 0}
+	WORLD_SIZE   = Size{Width: 350, Height: 550}
+)
