+++ debugger.go
@@ -5,10 +5,6 @@ import (
 	"image/color"
 	"time"
 
-	"github.com/eliasrenman/go-bullet-hell/assets"
-	"github.com/eliasrenman/go-bullet-hell/constant"
-	"github.com/eliasrenman/go-bullet-hell/entity"
-	"github.com/eliasrenman/go-bullet-hell/input"
 	"github.com/hajimehoshi/ebiten/v2"
 	"github.com/hajimehoshi/ebiten/v2/text"
 	"golang.org/x/image/font"
@@ -32,7 +28,7 @@ func NewDebugger(game *Game) *Debugger {
 		Visible: false,
 		Game:    game,
 
-		font:      assets.LoadFont("fonts/FiraCode.ttf", opentype.FaceOptions{}),
+		font:      LoadFont("fonts/FiraCode.ttf", opentype.FaceOptions{}),
 		fps:       0,
 		startTime: time.Now(),
 	}
@@ -40,7 +36,7 @@ func NewDebugger(game *Game) *Debugger {
 
 func (debugger *Debugger) Update() error {
 	// Toggle debug mode
-	if input.ButtonDebug.GetPressed(0) {
+	if ButtonDebug.GetPressed(0) {
 		debugger.Visible = !debugger.Visible
 		fmt.Printf("Debug mode: %v\n", debugger.Visible)
 	}
@@ -76,8 +72,8 @@ Total Game Objects: %v`,
 		debugger.startTime.Format("January 2 15:04:05"),
 		debugger.totalTime.Truncate(time.Second),
 		debugger.deltaTime.Truncate(time.Millisecond/100),
-		len(entity.GameObjects))
+		len(GameObjects))
 
-	text.Draw(screen, debugText, debugger.font, constant.PLAYFIELD_WIDTH+100, 5, color.White)
+	text.Draw(screen, debugText, debugger.font, PLAYFIELD_WIDTH+100, 5, color.White)
 
 }
