+++ game.go
@@ -1,38 +1,34 @@
 package main
 
 import (
-	"github.com/eliasrenman/go-bullet-hell/assets"
-	"github.com/eliasrenman/go-bullet-hell/constant"
-	"github.com/eliasrenman/go-bullet-hell/entity"
-	"github.com/eliasrenman/go-bullet-hell/geometry"
 	"github.com/hajimehoshi/ebiten/v2"
 )
 
 type Game struct {
-	player     *entity.Player
+	player     *Player
 	debugger   *Debugger
-	background assets.Background
+	background Background
 }
 
-var backgroundImage = assets.LoadImage("bg/img1.png", assets.OriginTopLeft)
+var backgroundImage = LoadImage("bg/img1.png", OriginTopLeft)
 
 func InitalizeGame() *Game {
-	player := entity.Spawn(entity.NewPlayer(geometry.Point{
-		X: constant.INITIAL_PLAYER_X,
-		Y: constant.INITIAL_PLAYER_Y,
+	player := Spawn(NewPlayer(Point{
+		X: INITIAL_PLAYER_X,
+		Y: INITIAL_PLAYER_Y,
 	}))
 
 	// Spawn boss
-	entity.Spawn(entity.NewBossOne(geometry.Point{
-		X: constant.INITIAL_PLAYER_X,
+	Spawn(NewBossOne(Point{
+		X: INITIAL_PLAYER_X,
 		Y: 200,
 	}))
 
 	game := Game{
 		player: player,
-		background: assets.Background{
+		background: Background{
 			Image:    backgroundImage,
-			Velocity: geometry.Up.ScaledBy(constant.STANDARD_BACKGROUND_SPEED),
+			Velocity: Up.ScaledBy(STANDARD_BACKGROUND_SPEED),
 		},
 		debugger: nil,
 	}
@@ -47,20 +43,20 @@ func (game *Game) Draw(screen *ebiten.Image) {
 	game.background.Draw(gameView)
 
 	// Draw game objects
-	entity.EachGameObject(func(obj entity.GameObject) {
+	EachGameObject(func(obj GameObject) {
 		obj.Draw(gameView)
 	})
 	DrawGameView(screen)
 	game.debugger.Draw(screen)
 }
 
-var gameView = ebiten.NewImage(constant.PLAYFIELD_WIDTH, constant.PLAYFIELD_HEIGHT)
+var gameView = ebiten.NewImage(PLAYFIELD_WIDTH, PLAYFIELD_HEIGHT)
 
 func DrawGameView(screen *ebiten.Image) {
 	op := &ebiten.DrawImageOptions{}
-	position := geometry.Point{X: float64(constant.PLAYFIELD_OFFSET), Y: float64(constant.PLAYFIELD_OFFSET)}
+	position := Point{X: float64(PLAYFIELD_OFFSET), Y: float64(PLAYFIELD_OFFSET)}
 
-	assets.TranslateScaleAndRotateImage(&op.GeoM, position, geometry.Size{Width: 1, Height: 1}, 0)
+	TranslateScaleAndRotateImage(&op.GeoM, position, Size{Width: 1, Height: 1}, 0)
 
 	screen.DrawImage(gameView, op)
 }
@@ -70,19 +66,19 @@ func (game *Game) Update() error {
 	game.background.Update()
 	game.debugger.Update()
 
-	entity.EachGameObject(func(obj entity.GameObject) {
+	EachGameObject(func(obj GameObject) {
 		obj.Update()
 	})
 
-	entity.SpawnGameObjects()
+	SpawnGameObjects()
 	return nil
 }
 
 func updateGameBackgroundSpeed(game *Game) {
 	if game.player.Velocity.Y != 0 {
-		offsetVelocity := ((game.player.Velocity.Y * -1) + constant.PLAYER_SPEED) / 2
-		game.background.Velocity = geometry.Up.ScaledBy(offsetVelocity + 1)
+		offsetVelocity := ((game.player.Velocity.Y * -1) + PLAYER_SPEED) / 2
+		game.background.Velocity = Up.ScaledBy(offsetVelocity + 1)
 	} else {
-		game.background.Velocity = geometry.Up.ScaledBy(constant.STANDARD_BACKGROUND_SPEED)
+		game.background.Velocity = Up.ScaledBy(STANDARD_BACKGROUND_SPEED)
 	}
 }
