+++ player.go
@@ -1,13 +1,8 @@
-package entity
+package main
 
 import (
 	"time"
 
-	"github.com/eliasrenman/go-bullet-hell/assets"
-	"github.com/eliasrenman/go-bullet-hell/constant"
-	"github.com/eliasrenman/go-bullet-hell/geometry"
-	"github.com/eliasrenman/go-bullet-hell/input"
-	"github.com/eliasrenman/go-bullet-hell/util"
 	"github.com/hajimehoshi/ebiten/v2"
 )
 
@@ -26,7 +21,7 @@ type Player struct {
 	DamageHitbox Collider
 }
 
-func NewPlayer(position geometry.Point) *Player {
+func NewPlayer(position Point) *Player {
 	player := &Player{
 		Entity: &Entity{
 			Position: position,
@@ -36,17 +31,17 @@ func NewPlayer(position geometry.Point) *Player {
 	}
 
 	player.MoveHitbox = &RectangleHitbox{
-		Size: geometry.Size{Width: 32, Height: 32},
+		Size: Size{Width: 32, Height: 32},
 		BaseHitbox: BaseHitbox{
-			Position: geometry.Vector{X: -16, Y: -16},
+			Position: Vector{X: -16, Y: -16},
 			Owner:    player.Entity,
 		},
 	}
 
 	player.DamageHitbox = &RectangleHitbox{
-		Size: geometry.Size{Width: 16, Height: 16},
+		Size: Size{Width: 16, Height: 16},
 		BaseHitbox: BaseHitbox{
-			Position: geometry.Vector{},
+			Position: Vector{},
 			Owner:    player.Entity,
 		},
 	}
@@ -58,21 +53,21 @@ func (player *Player) Start() {}
 
 var gameFieldHitbox = &RectangleHitbox{
 	BaseHitbox: BaseHitbox{
-		Position: geometry.Vector{X: 32, Y: 32},
+		Position: Vector{X: 32, Y: 32},
 	},
-	Size: geometry.Size{Width: constant.PLAYFIELD_WIDTH - 64, Height: constant.PLAYFIELD_HEIGHT - 64},
+	Size: Size{Width: PLAYFIELD_WIDTH - 64, Height: PLAYFIELD_HEIGHT - 64},
 }
 
 func (player *Player) Update() {
 	// Handle movement
-	moveInput := geometry.Vector{
-		X: input.AxisHorizontal.Get(0),
-		Y: -input.AxisVertical.Get(0),
+	moveInput := Vector{
+		X: AxisHorizontal.Get(0),
+		Y: -AxisVertical.Get(0),
 	}
 	direction := moveInput.Angle()
 	speed := 0.
 	if moveInput.X != 0 || moveInput.Y != 0 {
-		if input.ButtonSlow.Get(0) {
+		if ButtonSlow.Get(0) {
 			speed = moveSpeedSlow
 		} else {
 			speed = moveSpeed
@@ -80,8 +75,8 @@ func (player *Player) Update() {
 	}
 
 	// Allow sliding against walls
-	for i := 0.; i < 60 && i > -60; i = -(i + util.Sign(i)) {
-		mv := geometry.VectorFromAngle(direction + util.DegToRad(i)).ScaledBy(speed)
+	for i := 0.; i < 60 && i > -60; i = -(i + Sign(i)) {
+		mv := VectorFromAngle(direction + DegToRad(i)).ScaledBy(speed)
 		if player.MoveHitbox.CollidesAt(player.Position.Plus(mv), gameFieldHitbox) {
 			player.Velocity = mv
 			player.Move(mv)
@@ -90,11 +85,11 @@ func (player *Player) Update() {
 	}
 
 	// Handle shooting
-	if player.CanShoot && input.ButtonShoot.Get(0) {
+	if player.CanShoot && ButtonShoot.Get(0) {
 		if time.Since(player.lastShootTime) > time.Second/time.Duration(player.ShootSpeed) {
 			player.Shoot(
-				player.Position.Copy().Minus(geometry.Vector{X: 0, Y: 0}),
-				util.DegToRad(-90),
+				player.Position.Copy().Minus(Vector{X: 0, Y: 0}),
+				DegToRad(-90),
 				6,
 				25,
 			)
@@ -115,13 +110,13 @@ func (player *Player) Die() {
 }
 
 var (
-	playerImage      = assets.LoadImage("characters/player_forward.png", assets.OriginCenter)
-	playerLeftImage  = assets.LoadImage("characters/player_left.png", assets.OriginCenter)
-	playerRightImage = assets.LoadImage("characters/player_right.png", assets.OriginCenter)
+	playerImage      = LoadImage("characters/player_forward.png", OriginCenter)
+	playerLeftImage  = LoadImage("characters/player_left.png", OriginCenter)
+	playerRightImage = LoadImage("characters/player_right.png", OriginCenter)
 )
 
 func (player *Player) Draw(screen *ebiten.Image) {
-	hAxis := input.AxisHorizontal.Get(0)
+	hAxis := AxisHorizontal.Get(0)
 	image := playerImage
 
 	if hAxis < 0 {
@@ -130,5 +125,5 @@ func (player *Player) Draw(screen *ebiten.Image) {
 		image = playerRightImage
 	}
 
-	image.Draw(screen, player.Position, geometry.Size{Width: 1, Height: 1}, 0)
+	image.Draw(screen, player.Position, Size{Width: 1, Height: 1}, 0)
 }
