+++ rectanglehitbox.go
@@ -1,20 +1,19 @@
-package entity
+package main
 
 import (
 	"image/color"
 
-	"github.com/eliasrenman/go-bullet-hell/geometry"
 	"github.com/hajimehoshi/ebiten/v2"
 )
 
 type RectangleHitbox struct {
 	BaseHitbox
-	Size geometry.Size
+	Size Size
 }
 
-func (a *RectangleHitbox) CollidesWithRectangle(p geometry.Point, b *RectangleHitbox) bool {
+func (a *RectangleHitbox) CollidesWithRectangle(p Point, b *RectangleHitbox) bool {
 	aMin, aMax := a.GetRectangle(p)
-	bMin, bMax := b.GetRectangle(geometry.Vector{})
+	bMin, bMax := b.GetRectangle(Vector{})
 	if b.Owner != nil {
 		bMin, bMax = b.GetRectangle(b.Owner.Position)
 	}
@@ -22,11 +21,11 @@ func (a *RectangleHitbox) CollidesWithRectangle(p geometry.Point, b *RectangleHi
 	return aMin.X < bMax.X && aMax.X > bMin.X && aMin.Y < bMax.Y && aMax.Y > bMin.Y
 }
 
-func (a *RectangleHitbox) CollidesWithCircle(p geometry.Point, b *CircleHitbox) bool {
+func (a *RectangleHitbox) CollidesWithCircle(p Point, b *CircleHitbox) bool {
 	return b.CollidesWithRectangle(p, a)
 }
 
-func (hb *RectangleHitbox) GetRectangle(p geometry.Point) (geometry.Point, geometry.Point) {
+func (hb *RectangleHitbox) GetRectangle(p Point) (Point, Point) {
 	if hb.Owner == nil {
 		return hb.Position, hb.Position.Plus(hb.Size.AsVector())
 	}
@@ -39,7 +38,7 @@ func (hb *RectangleHitbox) CollidesWith(other Collider) bool {
 	return other.CollidesAt(hb.Owner.Position, hb)
 }
 
-func (a *RectangleHitbox) CollidesAt(p geometry.Point, b Collider) bool {
+func (a *RectangleHitbox) CollidesAt(p Point, b Collider) bool {
 	switch other := b.(type) {
 	case *RectangleHitbox:
 		return a.CollidesWithRectangle(p, other)
